rename the basis_func.py to basis_utils.py and refactoring after.
