+++ src/aspire/basis/pswf_2d.py
@@ -3,8 +3,8 @@ import numpy as np
 
 from aspire.basis import Basis
 from aspire.basis.pswf_utils import BNMatrix
-from aspire.basis.basis_func import d_decay_approx_fun, t_radial_part_mat, t_x_mat, t_x_derivative_mat
-from aspire.basis.basis_func import k_operator, leggauss_0_1
+from aspire.basis.basis_utils import d_decay_approx_fun, t_radial_part_mat, t_x_mat, t_x_derivative_mat
+from aspire.basis.basis_utils import k_operator, leggauss_0_1
 
 logger = logging.getLogger(__name__)
